requirements/base.txt
Django==2.1.7
djangorestframework==3.9.2

django-allauth==0.39.1
django-rest-auth==0.9.5
django-rest-knox==4.0.1

django-filter==2.1.0
django-webpack-loader==0.6.0

django-environ==0.4.5

whitenoise==4.1.2